+++ jest.config.js
@@ -4,6 +4,7 @@ module.exports = {
   testMatch: ['**/test/**/*.test.ts', '**/tests/**/*.test.ts'],
   collectCoverageFrom: ['src/**/*.ts'],
   coveragePathIgnorePatterns: ['/node_modules/', '/dist/'],
+  maxWorkers: 1,
   globals: {
     'ts-jest': {
       tsconfig: 'tsconfig.test.json',